Update patch set 1
Patch Set 1: Code-Review+1

The current CM-12.1 builds have a massive battery drain.
Due to some wakelocks the phone is unable to enter "Deep Sleep"
Battery drain is < 5% an hour in stand-by.
With "BetterBatteryStats" I found this wakelock to be one of the 2 causes which block "Deep Sleep"
"Wake Count" is almost as high as half of the total Wakelock counts.
This patch has been tested by me and resolves this wakelock.
Wakelock remaining is: "KeyEvents"

Patch-set: 1
Label: Code-Review=+1